Move file-input to input-screen/
